MKS937B simplify code by implementing code review comments
