--- docs/api/instruments/mksinst/adapters.rst
@@ -1,7 +1,0 @@
-########################
-MKS Instruments Adapters
-########################
-
-.. autoclass:: pymeasure.instruments.mksinst.adapters.MKSVISAAdapter
-    :members:
-    :show-inheritance:
