--- pymeasure/instruments/mksinst/adapters.py
@@ -1,56 +1,0 @@
-#
-# This file is part of the PyMeasure package.
-#
-# Copyright (c) 2013-2022 PyMeasure Developers
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-#
-
-from pymeasure.adapters import VISAAdapter
-
-
-class MKSVISAAdapter(VISAAdapter):
-
-    def __init__(self, resource_name, **kwargs):
-        kwargs.setdefault("write_termination", ";FF")
-        kwargs.setdefault("read_termination", ";")  # in reality its ";FF"
-        # which is, however, invalid for pyvisa. Therefore extra bytes have to
-        # be read in the read() method.
-        super().__init__(
-            resource_name,
-            **kwargs
-        )
-
-    def ask(self, command):
-        """ Writes the command to the instrument through the adapter
-        and returns the read response.
-        :param command: command string to be sent to the instrument
-        """
-        ret = super().ask(command)
-        t = super().read_bytes(2)  # read extra termination chars 'FF'
-        if t != b'FF':
-            raise ValueError(f"unexpected termination string received {t}")
-        return ret
-
-    def read(self):
-        ret = super().read()
-        t = super().read_bytes(2)  # read extra termination chars 'FF'
-        if t != b'FF':
-            raise ValueError(f"unexpected termination string received {t}")
-        return ret
